Introduce two args for Proba::operator()(real_t d, real_t l)
The arguments are
* `d`: time since infection
* `l`: additional parameter (living on the edge). Each Proba is free to
interpret as it wants. ConstantRate interprets as a Dt, i.e. the
duration of the contact (this is the most reasonable interpretation in
fact). For the moment and by default, if a Proba does not implement the
two-args operator(), then it is just a multiplative param to the
infectivity (note that this may bring the infectivity probability to a value > 1,
resulting in an error).